mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
package nextstep.mvc;

import jakarta.servlet.ServletException;
import jakarta.servlet.http.HttpServlet;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import nextstep.mvc.exception.NoSuchHandlerAdapterException;
import nextstep.mvc.exception.NoSuchRequestMappingException;
import nextstep.mvc.view.ModelAndView;
import nextstep.mvc.view.View;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.util.ArrayList;
import java.util.List;
import java.util.Objects;

public class DispatcherServlet extends HttpServlet {

    private static final long serialVersionUID = 1L;
    private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);

    private final List<HandlerMapping> handlerMappings;
    private final List<HandlerAdapter> handlerAdapters;

    public DispatcherServlet() {
        this.handlerMappings = new ArrayList<>();
        this.handlerAdapters = new ArrayList<>();
    }

    @Override
    public void init() {
        handlerMappings.forEach(HandlerMapping::initialize);
    }

    public void addHandlerMapping(HandlerMapping handlerMapping) {
        handlerMappings.add(handlerMapping);
    }

    public void addHandlerAdapters(HandlerAdapter handlerAdapter) {
        handlerAdapters.add(handlerAdapter);
    }

    @Override
    protected void service(HttpServletRequest request, HttpServletResponse response) throws ServletException {
        log.debug("Method : {}, Request URI : {}", request.getMethod(), request.getRequestURI());

        HandlerMapping handlerMapping = selectHandlerMapping(request);
        Object findHandler = handlerMapping.getHandler(request);
        HandlerAdapter handlerAdapter = selectHandlerAdapter(findHandler);

        try {
            ModelAndView modelAndView = handlerAdapter.handle(request, response, findHandler);
            View view = modelAndView.getView();
            view.render(modelAndView.getModel(), request, response);
        } catch (Exception e) {
            log.error("Exception : {}", e.getMessage(), e);
            throw new ServletException(e.getMessage());
        }
    }

    private HandlerMapping selectHandlerMapping(HttpServletRequest request) {
        return handlerMappings.stream()
                .filter(handlerMapping -> Objects.nonNull(handlerMapping.getHandler(request)))
                .findFirst().orElseThrow(() -> new NoSuchRequestMappingException("No such request mapping"));
    }

    private HandlerAdapter selectHandlerAdapter(Object findHandler) {
        return handlerAdapters.stream()
                .filter(handlerAdapter -> handlerAdapter.supports(findHandler))
                .findFirst().orElseThrow(() -> new NoSuchHandlerAdapterException("No such handler adapter"));
    }
}
